+++ docs/examples/tests/test_local_ray.py
@@ -13,6 +13,12 @@ import pytest
 
 from .parsers import get_snippet_as_str
 
+# Ray mishandles log file handlers and we get "_io.FileIO [closed]"
+# unraisable exceptions. Last tested with Ray 2.4.0.
+pytestmark = pytest.mark.filterwarnings(
+    "ignore::pytest.PytestUnraisableExceptionWarning"
+)
+
 
 class Snippets:
     @staticmethod
@@ -65,9 +71,6 @@ class TestSnippets:
 
     @staticmethod
     @pytest.mark.dependency()
-    # Ray mishandles log file handlers and we get "_io.FileIO [closed]"
-    # unraisable exceptions. Last tested with Ray 2.2.0.
-    @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
     def test_execute_workflow(change_test_dir, shared_ray_conn, change_db_location):
         # Given
 
@@ -90,9 +93,6 @@ class TestSnippets:
 
     @staticmethod
     @pytest.mark.dependency(depends=["TestSnippets::test_execute_workflow"])
-    # Ray mishandles log file handlers and we get "_io.FileIO [closed]"
-    # unraisable exceptions. Last tested with Ray 2.2.0.
-    @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
     def test_get_results(change_test_dir, shared_ray_conn, change_db_location):
         # Given