+++ src/orquestra/sdk/_ray/_client.py
@@ -8,9 +8,7 @@ import typing as t
 
 try:
     import ray
-    import ray._private.node
     import ray._private.ray_constants
-    import ray._private.utils
     import ray.runtime_env
     import ray.workflow
     from ray import exceptions  # noqa: F401
@@ -35,8 +33,6 @@ else:
     Storage = ray.workflow.storage.Storage
     RuntimeEnv = ray.runtime_env.RuntimeEnv
     FunctionNode = ray.dag.FunctionNode
-    SessionLatest = ray._private.node.SESSION_LATEST
-    TempDir = ray._private.utils.get_ray_temp_dir()
     LogPrefixActorName = ray._private.ray_constants.LOG_PREFIX_ACTOR_NAME
     LogPrefixTaskName = ray._private.ray_constants.LOG_PREFIX_TASK_NAME