+++ tests/runtime/api/test_api_with_ray.py
@@ -9,6 +9,12 @@ import pytest
 import orquestra.sdk as sdk
 from orquestra.sdk._base._testing import _connections
 
+# Ray mishandles log file handlers and we get "_io.FileIO [closed]"
+# unraisable exceptions. Last tested with Ray 2.4.0.
+pytestmark = pytest.mark.filterwarnings(
+    "ignore::pytest.PytestUnraisableExceptionWarning"
+)
+
 
 @sdk.task
 def sum_tuple_numbers(numbers: tuple):
@@ -40,9 +46,6 @@ class TestRunningLocalInBackground:
 
     class TestTwoStepForm:
         @staticmethod
-        # Ray mishandles log file handlers and we get "_io.FileIO [closed]"
-        # unraisable exceptions. Last tested with Ray 2.2.0.
-        @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
         def test_single_run(
             patch_config_location, ray, monkeypatch, tmp_path, mock_workflow_db_location
         ):
@@ -70,9 +73,6 @@ class TestRunningLocalInBackground:
 
     class TestReconnectToPreviousRun:
         @staticmethod
-        # Ray mishandles log file handlers and we get "_io.FileIO [closed]"
-        # unraisable exceptions. Last tested with Ray 2.2.0.
-        @pytest.mark.filterwarnings("ignore::pytest.PytestUnraisableExceptionWarning")
         def test_custom_save_locations(
             patch_config_location,
             ray,